update: assemble page to index, index to exampleindex
