--- src/pages/assemble.tsx
@@ -1,14 +1,0 @@
-import Layout from "@/components/Layout";
-import AssembleBoard from "@/features/ac_assemble/AssembleBoard";
-
-export default function Home() {
-  return (
-    <>
-      <Layout>
-        <section className="min-h-full py-40 w-full flex flex-col justify-center items-center gap-10">
-          <AssembleBoard />
-        </section>
-      </Layout>
-    </>
-  );
-}
